requirements.txt
Flask==2.0.1
numpy==1.19.5
tensorflow==2.4.1
Pillow==8.1.0
scikit-learn==0.24.2
opencv-python==4.5.1.48
imutils==0.5.4
matplotlib==3.3.4
